requirements.txt
numpy==1.21.5
scikit-learn==1.0.1
streamlit==1.5.0
langchain==0.2.2
python-dotenv==0.19.1
faiss-cpu==1.7.0